Source: https://github.com/msupply-foundation/open-msupply/commit/1a2462a6f7f2234a9be98a8653f7e11f2d68c172

+++ server/graphql/core/src/loader/item_stats.rs
@@ -12,7 +12,7 @@ pub struct ItemsStatsForItemLoader {
     pub service_provider: Data<ServiceProvider>,
 }
 
-pub type ItemStatsLoaderInputPayload = Option<u32>; // amc_lookback_months
+pub type ItemStatsLoaderInputPayload = Option<f64>; // amc_lookback_months
 pub type ItemStatsLoaderInput = IdPair<ItemStatsLoaderInputPayload>;
 impl ItemStatsLoaderInput {
     pub fn new(store_id: &str, item_id: &str, payload: ItemStatsLoaderInputPayload) -> Self {
